Extract duplicate test helper into a shared scope
