+++ FeedLoaderTests/Feed cache/CacheFeedUseCaseTests.swift
@@ -126,23 +126,4 @@ class CacheFeedUseCaseTests: XCTestCase {
         XCTAssertEqual(receivedError as NSError?, expectedError, file: file, line: line)
     }
     
-    private func uniqueItem() -> FeedImage {
-        return FeedImage(id: UUID(), description: "any", location: "any", url: anyUrl())
-    }
-    
-    private func uniqueImageFeed() -> (models: [FeedImage], localModels: [LocalFeedImage]) {
-        let items = [uniqueItem(), uniqueItem()]
-        let localItems = items.map { LocalFeedImage(id: $0.id, description: $0.description, location: $0.location, imageUrl: $0.url) }
-        return(items, localItems)
-    }
-    
-    private func anyUrl() -> URL {
-        return URL(string: "http://any-given-url.com")!
-    }
-    
-    private func anyError() -> NSError {
-        return NSError(domain: "any-error", code: 0, userInfo: nil)
-    }
-    
-    
 }
