+++ FeedLoaderTests/Feed cache/LoadFeedFromCacheUseCaseTests.swift
@@ -167,32 +167,5 @@ class LoadFeedFromCacheUseCaseTests: XCTestCase {
         wait(for: [expectation], timeout: 1.0)
         
     }
-
-    private func anyError() -> NSError {
-        return NSError(domain: "any-error", code: 0, userInfo: nil)
-    }
-    
-    private func uniqueItem() -> FeedImage {
-        return FeedImage(id: UUID(), description: "any", location: "any", url: anyUrl())
-    }
-    
-    private func anyUrl() -> URL {
-        return URL(string: "http://any-given-url.com")!
-    }
-    
-    private func uniqueImageFeed() -> (models: [FeedImage], localModels: [LocalFeedImage]) {
-        let items = [uniqueItem(), uniqueItem()]
-        let localItems = items.map { LocalFeedImage(id: $0.id, description: $0.description, location: $0.location, imageUrl: $0.url) }
-        return(items, localItems)
-    }
 }
 
-private extension Date {
-    func adding(days: Int) -> Date {
-        return Calendar(identifier: .gregorian).date(byAdding: .day, value: days, to: self)!
-    }
-    
-    func adding(seconds: TimeInterval) -> Date {
-        return self + seconds
-    }
-}
