+++ FeedLoaderTests/Feed cache/ValidateFeedCacheUseTestCaseFiles.swift
@@ -53,32 +53,6 @@ class ValidateFeedCacheUseTestCaseFiles: XCTestCase {
         trackMemoryLeak(instance: sut, file: file, line: line)
         return(store, sut)
     }
-    
-    private func anyError() -> NSError {
-        return NSError(domain: "any-error", code: 0, userInfo: nil)
-    }
-    
-    private func anyUrl() -> URL {
-        return URL(string: "http://any-given-url.com")!
-    }
-    
-    private func uniqueItem() -> FeedImage {
-        return FeedImage(id: UUID(), description: "any", location: "any", url: anyUrl())
-    }
-    
-    private func uniqueImageFeed() -> (models: [FeedImage], localModels: [LocalFeedImage]) {
-        let items = [uniqueItem(), uniqueItem()]
-        let localItems = items.map { LocalFeedImage(id: $0.id, description: $0.description, location: $0.location, imageUrl: $0.url) }
-        return(items, localItems)
-    }
 }
 
-private extension Date {
-    func adding(days: Int) -> Date {
-        return Calendar(identifier: .gregorian).date(byAdding: .day, value: days, to: self)!
-    }
-    
-    func adding(seconds: TimeInterval) -> Date {
-        return self + seconds
-    }
-}
+
